pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musae-inv"
version = "1.0.0"
description = "MuSAE-Inv: Multi-Layer Sparse Autoencoder with Invariant Causal Feature Selection for Cross-Domain Hallucination Detection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Vinayak Kumar"},
]
keywords = [
    "hallucination-detection",
    "sparse-autoencoders",
    "mechanistic-interpretability",
    "causal-feature-selection",
    "large-language-models",
    "cross-domain-generalization",
    "gemma",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24,<2.0",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "torch>=2.1",
    "transformers>=4.38",
    "datasets>=2.18",
    "accelerate>=0.29",
    "sae-lens>=4.0",
    "transformer_lens>=2.7",
    "einops>=0.7",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
vis = [
    "umap-learn>=0.5",
    "statsmodels>=0.14",
]
all = ["musae-inv[dev,vis]"]

[project.urls]
Homepage = "https://github.com/vvinayakkk/MuSAE-Inv-Invariant-Causal-Feature-Selection-from-Sparse"
Repository = "https://github.com/vvinayakkk/MuSAE-Inv-Invariant-Causal-Feature-Selection-from-Sparse"
Issues = "https://github.com/vvinayakkk/MuSAE-Inv-Invariant-Causal-Feature-Selection-from-Sparse/issues"

[project.scripts]
musae-train = "scripts.train:main"
musae-evaluate = "scripts.evaluate:main"
musae-extract = "scripts.extract_features:main"

[tool.setuptools.packages.find]
include = ["musae_inv*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
